requirements.txt
absl-py==2.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==24.10.0
bleach==6.0.0
blessed==1.20.0
blinker==1.8.2
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.1.0
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==1.2.0
fastapi==0.104.0
fastjsonschema==2.17.1
fastkml==0.12
filelock==3.12.4
firebase-admin==6.5.0
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.9.2
git-filter-repo==2.47.0
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.40
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
gspread==5.12.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.7
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirer==3.1.3
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
jedi==0.18.2
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.7
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.3
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-checkpoint-sqlite==2.0.1
langgraph-sdk==0.1.48
langsmith==0.1.147
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.15
mistune==3.0.1
ml_dtypes==0.5.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.2
onnxruntime==1.16.1
ooba==0.0.21
openai==1.59.7
opencv-contrib-python==4.10.0.84
opencv-python==4.8.1.78
opt_einsum==3.4.0
orjson==3.10.6
outcome==1.3.0
overrides==7.3.1
packaging==24.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.8.0
posthog==3.0.2
prometheus-client==0.17.0
prompt-toolkit==3.0.39
proto-plus==1.24.0
protobuf==4.25.5
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pygeoif==0.7
Pygments==2.15.1
PyJWT==2.9.0
pykml==0.2.0
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.1.2
PyPika==0.48.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-editor==1.0.4
python-http-client==3.3.7
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
readchar==4.0.5
redis==5.0.8
referencing==0.29.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.8.8
rq==1.16.2
rsa==4.9
scipy==1.14.1
selenium==4.14.0
Send2Trash==1.8.2
sendgrid==6.11.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.4.1
SQLAlchemy==2.0.36
stack-data==0.6.2
starkbank-ecdsa==2.2.0
starlette==0.27.0
sympy==1.12
tavily-python==0.3.5
tenacity==8.5.0
termcolor==2.3.0
terminado==0.17.1
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.14.1
tokentrim==0.1.10
tornado==6.3.2
tqdm==4.66.4
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==3.0.3
wget==3.2
widgetsnbextension==4.0.8
wsproto==1.2.0
yarl==1.9.2
yaspin==3.0.1
zipp==3.17.0
